requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.16
djangorestframework==3.12.1
idna==2.10
Pillow==7.2.0
python-decouple==3.3
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
stripe==2.45.0
urllib3==1.25.11
